requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dateutil>=2.8.2
